feat: atualização da regra de negócio das operações
