+++ src/main.c
@@ -18,44 +18,43 @@ int main() {
     do {
         printf("\n--- Menu de Gestao de Livros ---\n");
         printf("1. Inserir Livro\n");
-        printf("2. Buscar Livro\n");
+        printf("2. Buscar Livro (editar ou excluir)\n");
         printf("3. Imprimir Catalogo\n");
         printf("0. Sair\n");
         printf("Escolha uma opcao: ");
         scanf("%d", &opcao);
-        getchar();
+        getchar(); // Limpa o \n deixado por scanf
 
         switch (opcao) {
-        case 1: {
-            limpar_terminal();
-            Livro novoLivro = criarLivro();
-            NoBST *novo = novoNo(novoLivro, NULL, NULL);
-            catalogoPessoal = inserirNo(catalogoPessoal, novo);
-            break;
-        }
+            case 1: {
+                limpar_terminal();
+                Livro novoLivro = criarLivro();
+                NoBST *novo = novoNo(novoLivro, NULL, NULL);
+                catalogoPessoal = inserirNo(catalogoPessoal, novo);
+                break;
+            }
 
-        case 2:
-            limpar_terminal();
-            buscarLivro(catalogoPessoal);
-            break;
-        
-        case 3:
-            limpar_terminal();
-            printf("\n--- Catalogo de Livros ---\n");
-            imprimirCatalogo(catalogoPessoal);
-            break;
-        
-        case 0:
-            printf("Saindo...\n");
-            break;
+            case 2:
+                limpar_terminal();
+                buscarLivro(catalogoPessoal); 
+                break;
 
-        default:
-            printf("Opcao invalida!\n");
+            case 3:
+                limpar_terminal();
+                printf("\n--- Catalogo de Livros ---\n");
+                imprimirCatalogoCompleto(catalogoPessoal);
+                break;
+
+            case 0:
+                printf("Saindo...\n");
+                break;
+
+            default:
+                printf("Opcao invalida!\n");
         }
 
     } while (opcao != 0);
 
     liberarArvore(catalogoPessoal);
-
-return 0;
+    return 0;
 }
